replace spotbugs with quality; move controller code in to another package
