+++ buildSrc/build.gradle
@@ -10,11 +10,11 @@ repositories {
 }
 
 dependencies {
-    implementation 'com.diffplug.spotless:spotless-plugin-gradle:6.3.0'
+    implementation 'com.diffplug.spotless:spotless-plugin-gradle:6.5.1'
     implementation 'com.google.cloud.tools.jib:com.google.cloud.tools.jib.gradle.plugin:3.2.0'
     implementation 'com.srcclr.gradle:com.srcclr.gradle.gradle.plugin:3.1.10'
     implementation 'de.undercouch.download:de.undercouch.download.gradle.plugin:5.0.1'
-    implementation 'gradle.plugin.com.github.spotbugs.snom:spotbugs-gradle-plugin:4.7.2'
+    implementation 'ru.vyarus:gradle-quality-plugin:4.7.0'
     implementation 'io.spring.dependency-management:io.spring.dependency-management.gradle.plugin:1.0.11.RELEASE'
     implementation 'org.hidetake.swagger.generator:org.hidetake.swagger.generator.gradle.plugin:2.19.1'
     implementation 'org.sonarqube:org.sonarqube.gradle.plugin:3.3'
