+++ buildSrc/src/main/groovy/bio.terra.catalog.java-common-conventions.gradle
@@ -4,7 +4,6 @@ plugins {
     id 'java'
 
     id 'com.diffplug.spotless'
-    id 'com.github.spotbugs'
     id 'org.hidetake.swagger.generator'
 }
 
@@ -74,21 +73,6 @@ compileJava {
         dependsOn(spotlessCheck)
     } else {
         dependsOn(spotlessApply)
-    }
-}
-
-// Spotbugs configuration
-spotbugs {
-    reportLevel = 'high'
-    effort = 'max'
-}
-spotbugsMain {
-    reports {
-        if (isCiServer) {
-            sarif.enabled = true
-        } else {
-            html.enabled = true
-        }
     }
 }
