+++ service/build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'com.google.cloud.tools.jib'
     id 'com.srcclr.gradle'
     id 'org.sonarqube'
+    id 'ru.vyarus.quality'
 
     id 'com.gorylenko.gradle-git-properties' version '2.3.1'
     id 'org.liquibase.gradle' version '2.1.0'
@@ -58,4 +59,10 @@ liquibase {
             logLevel 'info'
         }
     }
+}
+
+quality {
+    checkstyle false
+    pmd false
+    exclude = ['bio/terra/catalog/model/**', 'bio/terra/catalog/api/**']
 }
