+++ service/src/test/java/bio/terra/catalog/api/DatasetApiControllerTest.java
@@ -14,6 +14,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import bio.terra.catalog.common.StorageSystem;
+import bio.terra.catalog.controller.DatasetApiController;
+import bio.terra.catalog.controller.GlobalExceptionHandler;
 import bio.terra.catalog.model.ColumnModel;
 import bio.terra.catalog.model.CreateDatasetRequest;
 import bio.terra.catalog.model.DatasetPreviewTable;
