+++ service/src/test/java/bio/terra/catalog/api/PublicApiControllerTest.java
@@ -8,6 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import bio.terra.catalog.config.VersionConfiguration;
+import bio.terra.catalog.controller.PublicApiController;
 import bio.terra.catalog.model.SystemStatus;
 import bio.terra.catalog.service.CatalogStatusService;
 import org.junit.jupiter.api.Test;
